MAINT: stats._ProbabilityDistribution: move to separate file to simplify diff
